tests/wpt/web-platform-tests/service-workers/service-worker/request-end-to-end.https.html

<!DOCTYPE html>
<title>Service Worker: Request end-to-end</title>
<script src="/resources/testharness.js"></script>
<script src="/resources/testharnessreport.js"></script>
<script src="resources/test-helpers.sub.js"></script>
<script>
var t = async_test('Request: end-to-end');
t.step(function() {
    var url = 'resources/request-end-to-end-worker.js';
    var scope = 'resources/blank.html';
    var frames = [];

    service_worker_unregister_and_register(t, url, scope)
      .then(onRegister)
      .catch(unreached_rejection(t));

    function sendMessagePort(worker) {
        var messageChannel = new MessageChannel();
        worker.postMessage({port:messageChannel.port2}, [messageChannel.port2]);
        return messageChannel.port1;
    }

    function onRegister(registration) {
        var sw = registration.installing;
        var port = sendMessagePort(sw);
        port.addEventListener('message', t.step_func(function(event) {
            onMessage(event);
        }), false);
        port.start();
        sw.addEventListener('statechange', t.step_func(function(event) {
            if (event.target.state == 'activated')
                onActive();
        }));
    }

    function onActive() {
        with_iframe(scope).then(function(f) { frames.push(f); });
    }

    function onMessage(event) {
        assert_equals(
            event.data.url,
            location.href.substring(0, location.href.lastIndexOf('/') + 1) +
            scope,
            'request.url should be passed to onfetch event.');
        assert_equals(event.data.mode, 'navigate',
                      'request.mode should be passed to onfetch event.');
        assert_equals(event.data.method, 'GET',
                      'request.method should be passed to onfetch event.');
        assert_equals(event.data.referrer, location.href,
                      'request.referrer should be passed to onfetch event.');
        assert_equals(event.data.headers['user-agent'], undefined,
                      'Default User-Agent header should not be passed to onfetch event.')
        assert_equals(event.data.errorNameWhileAppendingHeader, 'TypeError',
                      'Appending a new header to the request must throw a ' +
                      'TypeError.')
        frames.forEach(function(f) { f.remove(); });
        service_worker_unregister_and_done(t, scope);
    }
});
</script>
